Update evaluation.markdown
The patch to the course is focused on the evaluation. The patch component has been removed and specific amounts assigned to the presentation and proposal for the research paper with an increase in the value assigned to the memos and less assigned to class participation, but a separate amount assigned to the team reports with an increase to 2 or 3 reports.